feat(gateway): replace Full-NAT edgenat.c with Maglev/DSR edgedsr.c
Removes internal/plumbing/ebpf/edgeprog/edgenat.c (the Full-NAT DNAT+SNAT
XDP datapath) entirely and replaces it with edgedsr.c: a Maglev-style
consistent-hash forward-only datapath that does no address or port
rewriting at all (design plan §0). Breaking change, no migration path, no
Full-NAT mode retained alongside it.

What this removes, and why it's now unnecessary:
- conn_table (BPF_MAP_TYPE_LRU_HASH, 65536 entries) -- Full-NAT needed it
  to remember a flow's assigned backend/SNAT port; DSR has no per-flow
  state to remember, since the same Maglev table produces the same
  backend for the same flow on every packet.
- The return/decap branch (handle_return) -- existed solely to route
  Full-NAT's own SNAT'd replies back through this node. DSR never sees
  replies (the backend answers the client directly), so there is nothing
  to decap here.
- PAT/SNAT-port allocation and all checksum-fixup logic (bpf_csum_diff,
  csum_fold_add) -- the packet is forwarded completely unmodified, so its
  own checksum is already correct.

What's kept, unchanged: push_outer_header/resolve_fib_and_write_eth's
SRv6 encap-push mechanics (grow-by-40, write a fresh Ethernet+IPv6 header,
resolve the L2 next-hop via bpf_fib_lookup) -- encapsulating toward a
backend's worker-node uSID is identical work with or without NAT.

New vip_table (replaces rule_table) adds a precomputed Maglev lookup table
per VIP (EDGE_MAGLEV_TABLE_SIZE=1021, prime) alongside the existing
backend list -- built by the Go control plane's internal/maglev.Table, not
computed in-kernel. EDGE_MAX_BACKENDS raised from edgenat.c's 8 to 64,
matching NetworkRuleSpec.Backends' own actual CRD limit (a pre-existing
mismatch, closed as a side effect of this rewrite rather than carried
forward unexamined).

Validated against a real kernel via BPF_PROG_TEST_RUN
(edgedsr_test.go): confirmed the core DSR property directly, not just by
inspection -- a claimed VIP's packet is pushed toward the backend's uSID
with the inner packet byte-for-byte identical to the client's original.
One real verifier gotcha found and fixed this way: unlike a literal array
index, the verifier does not track a bounded range through clang's
emitted division-by-constant sequence for hash % 1021 (1021 is prime, so
clang lowers the mod to a multiply/shift reciprocal trick, not a mask) --
needs the same EDGE_BARRIER_VAR-then-clamp treatment as the pre-existing
backend-index bounds problem, just applied as a range clamp instead of a
bitmask (1021 isn't a power of two).

Still to come (separate, following commits): internal/gateway,
internal/plumbing/ebpf/edgemap, internal/controller's
NetworkGateway/NetworkRule reconcilers, internal/plumbing/ebpf/edgemetrics,
and cmd/galactic-gateway all still reference the removed edgenat.c types
and will not build until updated for edgedsr.c's new shapes.

Part of the DSR/Maglev gateway redesign (kitt notebook
projects/galactic/dsr-maglev-nptv6-nat66-design.md).
