+++ internal/plumbing/ebpf/edgeprog/doc.go
@@ -3,21 +3,27 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 // Package edgeprog holds the compiled XDP program that implements the edge
-// gateway's ingress Full-NAT (DNAT+SNAT) load-balancing and reverse un-NAT
-// datapath, plus a direct SRv6 uSID push/strip -- no Geneve overlay, no
-// separate gateway-tier decap step, no kernel VRF/FIB dependency for the
-// encap itself (see the design plan at
-// /home/sprygada/.claude/datum/plans/merry-percolating-tulip.md for the
-// full rationale versus the earlier, rejected gwprog/Geneve approach).
+// gateway's Maglev/DSR (Direct Server Return) consistent-hash
+// load-balancing datapath, plus a direct SRv6 uSID push -- no Geneve
+// overlay, no separate gateway-tier decap step, no kernel VRF/FIB
+// dependency for the encap itself, and (unlike the Full-NAT predecessor
+// this replaces) no address/port rewriting or reverse-path decap of any
+// kind: the backend answers the client directly, so this program only
+// ever needs a forward path. See the DSR/Maglev design plan (kitt
+// notebook, projects/galactic/dsr-maglev-nptv6-nat66-design.md) for the
+// full rationale, and edgedsr.c's own header comment for what this
+// simplification removes versus the Full-NAT edgenat.c it replaces
+// (removed entirely, not kept as an alternate mode -- breaking change, no
+// migration path).
 //
-// edgenat.c is the single source of truth for the packet path; see its
+// edgedsr.c is the single source of truth for the packet path; see its
 // header comment for the full walkthrough. `go generate` (via bpf2go,
 // github.com/cilium/ebpf's code generator) compiles it with clang into a
 // CO-RE-portable BPF object and generates matching Go bindings
-// (EdgenatObjects, LoadEdgenat, LoadEdgenatObjects, plus per-map/
+// (EdgedsrObjects, LoadEdgedsr, LoadEdgedsrObjects, plus per-map/
 // per-program fields) in this package -- run `go generate ./...` from the
 // repo root, or `go generate` from this directory, after editing
-// edgenat.c.
+// edgedsr.c.
 //
 // Same not-committed-generated-artifacts convention as
 // internal/plumbing/ebpf/prog (see that package's doc.go): *_bpfel.go/
@@ -42,14 +48,8 @@ package edgeprog
 // multiarch directories and why -cc is omitted (same clang/bpf2go
 // environment, same rationale, not repeated here).
 //
-// -Wno-address-of-packed-member suppresses a known clang false positive
-// for this file's l4_view.{sport,dport,check}_ptr fields: taking the
-// address of a __be16 member inside a packed struct warns generically
-// (packed structs can place a field at any byte offset), but every field
-// edgenat.c takes the address of sits at a compile-time-fixed *even*
-// offset (edge_tcphdr's source/dest/check are 0/2/16; edge_udphdr's are
-// 0/2/6) -- genuinely 2-byte aligned despite the struct's packed
-// attribute, so there is no real unaligned-access risk to suppress
-// unsafely here.
+// Unlike edgenat.c, edgedsr.c never takes the address of a field inside a
+// packed struct (no rewrite-by-pointer of any header field -- see its own
+// header comment), so -Wno-address-of-packed-member is no longer needed.
 //
-//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -cflags "-O2 -g -Wall -Wno-address-of-packed-member -idirafter /usr/include/x86_64-linux-gnu -idirafter /usr/include/aarch64-linux-gnu" -target bpfel,bpfeb -type rule_key -type backend -type rule_value -type rule_stats_value -type conn_key -type conn_value -type gw_config Edgenat edgenat.c
+//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -cflags "-O2 -g -Wall -idirafter /usr/include/x86_64-linux-gnu -idirafter /usr/include/aarch64-linux-gnu" -target bpfel,bpfeb -type vip_key -type backend -type vip_value -type vip_stats_value -type encap_config Edgedsr edgedsr.c
